fix(file-size): move AuditCtx to gemini_audit module to stay under 1000-line cap
The previous commit (`refactor(gemini_client): replace AuditCtx::Default
with explicit no_audit()`) added the `impl AuditCtx { pub fn no_audit }`
block that pushed gemini_client.rs to 1012 lines — over the airuleset
1000-line cap, failing the File Size Check on PR CI.

Move AuditCtx to gemini_audit.rs (its natural home — it's the *context*
for audit writes). Re-export from gemini_client.rs so existing
`use crate::lyrics::gemini_client::AuditCtx;` imports keep compiling.

gemini_client.rs: 1012 → 989 lines.
gemini_audit.rs: 322 → 350 lines.
